fix(review): address PR #16 review feedback
- initiative_handler.go: normalize sort_by/sort_dir with strings.ToLower
  so "ASC"/"DESC" and "Supporters" work the same as lowercase (per copilot-pull-request-reviewer)
- initiative_repository.go: add i.created_on/i.id secondary sort for
  deterministic pagination when sorting by supporters/total_raised (per copilot-pull-request-reviewer)
- initiative_repository.go: cast SUM(amount_in_cents) to ::bigint to
  avoid pgx scan failure on NUMERIC result (per copilot-pull-request-reviewer)
- initiative_repository.go: initialize Goals as empty slice so JSON
  serializes as [] not null (per copilot-pull-request-reviewer)
- statistics_repository.go: cast SUM aggregates to ::bigint to avoid
  pgx scan failure on NUMERIC result (per copilot-pull-request-reviewer)
- 001_initial.up.sql: add DESC indexes on initiative_ledger_stats for
  supporters and total_raised_cents sort columns (per copilot-pull-request-reviewer)

+++ backend/internal/infrastructure/db/statistics_repository.go
@@ -32,9 +32,9 @@ func (r *StatisticsRepository) GetPlatformStatistics(ctx context.Context) (*mode
 
 	const q = `
 		SELECT
-			COALESCE(SUM(ls.total_raised_cents), 0) AS total_raised_cents,
-			COALESCE(SUM(ls.supporters), 0)         AS total_supporters,
-			COUNT(i.id)                             AS total_initiatives
+			COALESCE(SUM(ls.total_raised_cents), 0)::bigint AS total_raised_cents,
+			COALESCE(SUM(ls.supporters), 0)::bigint         AS total_supporters,
+			COUNT(i.id)::bigint                             AS total_initiatives
 		FROM initiatives i
 		INNER JOIN initiative_ledger_stats ls ON ls.initiative_id = i.id
 		WHERE i.status = 'published'`
